[REFACTOR] domain 모듈 리팩토링
