+++ smeem-domain/src/main/java/com/smeem/domain/diary/repository/DiaryRepository.java
@@ -1,15 +1,9 @@
 package com.smeem.domain.diary.repository;
 
-
 import com.smeem.domain.diary.model.Diary;
 import com.smeem.domain.member.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
-
     void deleteAllByMember(Member member);
-
-    List<Diary> findAllByMemberId(Long memberId);
 }
