+++ smeem-domain/src/main/java/com/smeem/domain/topic/repository/TopicCustomRepository.java
@@ -1,6 +1,5 @@
 package com.smeem.domain.topic.repository;
 
-
 import com.smeem.domain.topic.model.Topic;
 
 public interface TopicCustomRepository {
